Move test from `MempoolMirrorTests` to `MempoolTests` class
